+++ src/index.html
@@ -19,7 +19,7 @@
         </div>
         <hr />
         <div>
-          <section id="todo-form">
+          <form action="" method="post" id="todo-form">
             <div class="todo-edit flex">
               <input
                 type="text"
@@ -31,9 +31,9 @@
             <hr />
             <div class="todo-inner"></div>
             <div class="clear-all">
-              <h3 class="clear">Clear all Completed</h3>
+              <h3>Clear all Completed</h3>
             </div>
-          </section>
+          </form>
         </div>
       </div>
     </main>
